Changes recovery system initialization.
Changes the initialization. Previously, if memory, barometer or actuator do not initialize properly, the recovery system became blocked. Now, if barometer or actuator do not initialize properly, another call to the initialization method is performed after a delay.  The execution is blocked until the barometer and actuator is initialized properly.